Create Interactive Reporting with Drilldown
In this project, I have enhanced the implementation of a classic reporting task. The primary modification allows the user to sort the list based on different criteria using the function keys <F4>, <F5>, or <F6>. Initially, an unsorted list is displayed, and subsequent presses of the function keys create branching lists sorted by price, flight date, or occupancy. If a key is pressed multiple times (e.g., twice for <F4>), the sorting order switches between ascending and descending.

To inform the user about available sorting options and the current sorting criterion, I utilized the TOP-OF-PAGE and TOP-OF-PAGE DURING LINE-SELECTION events.

Additionally, I extended the report with interactive functionality.  Now, a double-click on a displayed flight allows the user to drill down and view the list of bookings, including customer names and addresses from the SCUSTOM table. I employed appropriate timing instructions to keep the user informed about the content displayed on the current list.

The relevant tables for data retrieval are SCARR, SPFLI, SFLIGHT, SBOOK, and SCUSTOM.